Updated EXP data reading
The total can go over int.Max, so we can't treat it as an int. Instead it should be a long, and we'll adjust the value sent to the client as appropriate, like in ZC_MAX_EXP_CHANGED, where the total EXP are sent.

+++ src/ChannelServer/World/Character.cs
@@ -313,7 +313,7 @@ namespace Melia.Channel.World
 		/// <summary>
 		/// Total number of accumulated experience points.
 		/// </summary>
-		public int TotalExp { get; set; }
+		public long TotalExp { get; set; }
 
 		/// <summary>
 		/// Hp multiplicator from the basic job.
